Some minor cleanup.

+++ src/emu/logical/cmp.rs
@@ -16,15 +16,16 @@ pub fn execute_cmp(state: &mut State, cmp: CMP) {
             let lhs = state.get_register(&Register::A) as u16;
             let rhs = state.get_register(&r) as u16;
 
-            let result = lhs.wrapping_sub(rhs) as u16;
+            let result = lhs.wrapping_sub(rhs);
 
             // The accumulator remains unchanged.
             state.alu.flags = Flags {
-                zero: z_flag!(result & 0x00FF),
+                zero: z_flag!((result & 0x00FF) as u8),
                 carry: result >> 8 != 0,
-                sign: s_flag!(result & 0x00FF),
-                parity: p_flag!(result & 0x00FF),
-                auxiliary_carry: ((state.alu.accumulator as u16 ^ result ^ rhs).not() & 0x0010) != 0,
+                sign: s_flag!((result & 0x00FF) as u8),
+                parity: p_flag!((result & 0x00FF) as u8),
+                auxiliary_carry: ((state.alu.accumulator as u16 ^ result ^ rhs).not() & 0x0010)
+                    != 0,
             };
         }
         CMP::CompareMemory => {
@@ -32,15 +33,16 @@ pub fn execute_cmp(state: &mut State, cmp: CMP) {
             let lhs = state.get_register(&Register::A) as u16;
             let rhs = state.get_memory(address) as u16;
 
-            let result = lhs.wrapping_sub(rhs) as u16;
+            let result = lhs.wrapping_sub(rhs);
 
             // The accumulator remains unchanged.
             state.alu.flags = Flags {
-                zero: z_flag!(result & 0x00FF),
+                zero: z_flag!((result & 0x00FF) as u8),
                 carry: result >> 8 != 0,
-                sign: s_flag!(result & 0x00FF),
-                parity: p_flag!(result & 0x00FF),
-                auxiliary_carry: ((state.alu.accumulator as u16 ^ result ^ rhs).not() & 0x0010) != 0,
+                sign: s_flag!((result & 0x00FF) as u8),
+                parity: p_flag!((result & 0x00FF) as u8),
+                auxiliary_carry: ((state.alu.accumulator as u16 ^ result ^ rhs).not() & 0x0010)
+                    != 0,
             };
         }
     }
