+++ src/emu/logical/cpi.rs
@@ -13,15 +13,16 @@ pub fn execute_cpi(state: &mut State, cpi: CPI) {
             let lhs = state.get_register(&Register::A) as u16;
             let rhs = data as u16;
 
-            let result = lhs.wrapping_sub(rhs) as u16;
+            let result = lhs.wrapping_sub(rhs);
 
             // The accumulator remains unchanged.
             state.alu.flags = Flags {
-                zero: z_flag!(result & 0x00FF),
+                zero: z_flag!((result & 0x00FF) as u8),
                 carry: result >> 8 != 0,
-                sign: s_flag!(result & 0x00FF),
-                parity: p_flag!(result & 0x00FF),
-                auxiliary_carry: ((state.alu.accumulator as u16 ^ result ^ rhs).not() & 0x0010) != 0,
+                sign: s_flag!((result & 0x00FF) as u8),
+                parity: p_flag!((result & 0x00FF) as u8),
+                auxiliary_carry: ((state.alu.accumulator as u16 ^ result ^ rhs).not() & 0x0010)
+                    != 0,
             };
         }
     }
